+++ examples/vite-swr/.wundergraph/wundergraph.config.ts
@@ -22,16 +22,22 @@ configureWunderGraphApplication({
 	codeGenerators: [
 		{
 			templates: [...templates.typescript.all],
-			path: './generated',
 		},
 		{
 			templates: [templates.typescript.client],
-			path: '../components/generated',
+			path: '../src/generated',
 		},
 	],
 	cors: {
 		...cors.allowAll,
-		allowedOrigins: ['http://localhost:5173'],
+		allowedOrigins: process.env.NODE_ENV === 'production' ? ['https://*'] : ['http://*'],
+		/**
+		 * Please configure CORS carefully to make sure that your users are protected.
+		 * Allowing all origins is usually the worst possible configuration.
+		 *
+		 * @docs https://docs.wundergraph.com/docs/wundergraph-config-ts-reference/configure-cors
+		 */
+		// allowedOrigins: process.env.NODE_ENV === 'production' ? ['http://your.app'] : ['http://localhost:3000'],
 	},
 	authentication: {
 		cookieBased: {