+++ packages/sdk/src/client/client.ts
@@ -18,13 +18,36 @@ import { serialize } from '../utils';
 import { GraphQLResponseError } from './GraphQLResponseError';
 import { ResponseError } from './ResponseError';
 import { InputValidationError } from './InputValidationError';
+import { AuthorizationError } from './AuthorizationError';
+import { ClientResponseError } from './ClientResponseError';
 
 // https://graphql.org/learn/serving-over-http/
 
 export interface UploadValidationOptions {
+	/** Whether authentication is required to upload to this profile
+	 *
+	 * @default true
+	 */
+	requireAuthentication?: boolean;
+	/** Maximum file size allowed per upload
+	 *
+	 * @default 10 * 1024 * 1024 (10MB)
+	 */
 	maxAllowedUploadSizeBytes?: number;
+	/** Maximum number of files allowed per upload
+	 *
+	 * @default unlimited
+	 */
 	maxAllowedFiles?: number;
+	/** List of allowed file extensions
+	 *
+	 * @default Any
+	 */
 	allowedFileExtensions?: string[];
+	/** List of allowed mime types
+	 *
+	 * @default Any
+	 */
 	allowedMimeTypes?: string[];
 }
 
@@ -119,23 +142,21 @@ export class Client {
 	}
 
 	// Determines whether the body is unparseable, plain text, or json (and assumes an invalid input if json)
-	private async handleClientResponseError(response: globalThis.Response): Promise<ClientResponse> {
+	private async handleClientResponseError(response: globalThis.Response): Promise<ClientResponseError> {
 		const text = await response.text();
-		if (!text) {
-			return {
-				error: new ResponseError('Unable to parse response body', response.status),
-			};
-		}
 		try {
 			const json = JSON.parse(text);
-			return {
-				error: new InputValidationError(json, response.status),
-			};
+
+			switch (response.status) {
+				case 401:
+					return new AuthorizationError(json.errors[0]?.message, response.status);
+				case 400:
+					return new InputValidationError(json, response.status);
+				default:
+					return new ResponseError(json.errors[0]?.message || json.message, response.status);
+			}
 		} catch {
-			// if the JSON.parse fails, we know the body must be plaintext
-			return {
-				error: new ResponseError(text, response.status),
-			};
+			return new ResponseError(text.length ? text : 'Response is not OK', response.status);
 		}
 	}
 
@@ -149,7 +170,7 @@ export class Client {
 		// even if the response is an HTTP 404 or 500.
 
 		if (!response.ok) {
-			return this.handleClientResponseError(response);
+			return { error: await this.handleClientResponseError(response) };
 		}
 
 		const json = await response.json();
@@ -403,7 +424,7 @@ export class Client {
 
 		const headers: Headers = {};
 
-		if (this.csrfEnabled) {
+		if (this.csrfEnabled && (validation?.requireAuthentication ?? true)) {
 			headers['X-CSRF-Token'] = await this.getCSRFToken();
 		}