+++ pkg/node/node.go
@@ -9,7 +9,6 @@ import (
 	"net"
 	"net/http"
 	"os"
-	"strings"
 	"time"
 
 	"github.com/gorilla/mux"
@@ -423,9 +422,7 @@ func (n *Node) startServer(nodeConfig WunderNodeConfig) error {
 		TLSHandshakeTimeout: 10 * time.Second,
 	}
 
-	serverUrl := strings.TrimSuffix(nodeConfig.Api.Options.ServerUrl, "/")
-
-	hooksClient := hooks.NewClient(serverUrl, n.log)
+	hooksClient := hooks.NewClient(nodeConfig.Api.Options.ServerUrl, n.log)
 
 	transportFactory := apihandler.NewApiTransportFactory(nodeConfig.Api, hooksClient, n.options.enableDebugMode)
 
@@ -448,12 +445,11 @@ func (n *Node) startServer(nodeConfig WunderNodeConfig) error {
 		EnableIntrospection:        n.options.enableIntrospection,
 		GitHubAuthDemoClientID:     n.options.githubAuthDemo.ClientID,
 		GitHubAuthDemoClientSecret: n.options.githubAuthDemo.ClientSecret,
-		HookServerURL:              serverUrl,
 		DevMode:                    n.options.devMode,
 	}
 
 	n.builder = apihandler.NewBuilder(n.pool, n.log, loader, hooksClient, builderConfig)
-	internalBuilder := apihandler.NewInternalBuilder(n.pool, n.log, loader)
+	internalBuilder := apihandler.NewInternalBuilder(n.pool, n.log, hooksClient, loader)
 
 	publicClosers, err := n.builder.BuildAndMountApiHandler(n.ctx, router, nodeConfig.Api)
 	if err != nil {